+++ src/control_plane/desired_state/mod.rs
@@ -1,9 +1,11 @@
 mod desired_project;
 mod desired_project_error;
 mod desired_service;
+mod desired_service_options;
 mod resolve_desired_project;
 
 pub(crate) use desired_project::DesiredProject;
 pub(crate) use desired_project_error::DesiredProjectError;
 pub(crate) use desired_service::DesiredService;
+pub(crate) use desired_service_options::DesiredServiceOptions;
 pub(crate) use resolve_desired_project::resolve_desired_project;